Renames _utils.py -> utils.py
